requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pgvector
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
alembic
celery
redis
httpx
PyPDF2
python-docx
openai
ollama
loguru
pytest
pytest-asyncio
